Delete 3.1
Suggest deleting 3.1. "You may use and remix the Community marks on the Community projects as you please." If the "project" is something distributable, like software, then you are telling people it is ok to modify product and then use a bastardized trademark on it. That is harmful to the brand and possibly an express naked license. I don't know in what context this provision makes sense except for something non-distributable, like Wikipedia.