[feature] change social media account name default
[This is probably my first pull request intended to create a substantive policy change, rather than merely clarifying existing policies as I understand them. Triage/ignore/close with that in mind.]

I think for most projects, social network names like "@ILoveProjectName" are pretty appropriate and non-confusing, especially when paired with a clear statement in the social network bio field. (See, for example, @NYTIsOnIt's bio.) This proposed language reflects that as a default.

If we had an implementer's guide (which I am slowly working on in a branch) it might be appropriate to pair this with a comment effectively saying "remove this sentence if for some reason your brand is particularly weak or easy to confuse".