[feature] explicitly permit attribution
[Substantive policy change, rather than mere clarification. Triage/ignore/close with that in mind.]

This is probably mostly implied by 3.5.4, but I think it is useful to make clear that attribution in compliance with the community's copyright license is permitted.

Optionally, could consider adding an explicit statement that requires compliance with any requirements to indicate change like those in GPL v2 Sec. 2(a) or CC BY-SA 4.0 Sec. 3(a)(1)(B).

We could