[feature] call out non-confusion
[Could be interpreted as a substantive policy change, rather than mere clarification. Triage/ignore/close with that in mind.]

The last sentence of Sec. 2.2 makes an important point - your use can't suggest endorsement or affiliation.  I think it is worth pulling out from Sec. 2.2 and making clear that it applies to all uses, regardless of notice, trademark symbol, etc. This change suggests language for that point.

I considered adding ", but may not be enough, depending on what situation you're using the mark in" at the end of the second sentence, but I think maybe that obfuscates more than it illuminates so left it out in this pull request.

(Comparatively, I note that the Model Trademark Guidelines leads with non-confusion, and the more I think about it, the more I like that idea, but down that road lies a significant rewrite, which is exactly what I'm trying to avoid here so that you can go ahead and accept my pull requests ;)