Globalize nominative use language
This phrase was somewhat US-centric, which makes sense for WMF but not for the template. This change:
1. Removes the reference to US law; and
2. Adds a reference to "referential" use, which is the preferred nomenclature in the EU (see, e.g., INTA on EU relevant statutory provisions: http://www.inta.org/INTABulletin/Pages/EU_TM_Reform_3_7022.aspx )

This is not perfect, as the caselaw in the EU is not as favorable to third parties as in the US, but Article 12 of the harmonized EU regulation does appear to have similar intent as US nominative fair use.

If we wanted to be more precise, one might drop "fair" here (it isn't terribly well-defined in US trademark law) and simply stick to nominative/referential. But that might be confusing to non-legal readers.